Add additional test cases to GitStatusTesterTest for better coverage of file status detection
